+++ task1/internal/library/Structures.go
@@ -1,10 +1,6 @@
 package library
 
-import (
-	"task1/mylibrary/internal/storage"
-)
-
 type library struct {
-	storage storage.Storage
+	storage Storage
 	idGen   func(title string) uint32
 }
